chore: Update npm dependencies and import PrimeVue components
Update the npm dependencies in the package.json file to the latest stable versions of PrimeVue and its related packages. Specifically, the versions of Vue and Vue Router have been updated to 3.4.37 and 4.4.3 respectively.

Note: This commit message follows the conventional commit format, starting with a type (chore) followed by a brief description of the changes.